created much more thorough embedded documentation
rational: increase usability and readability

While the interface of this module is very Pythonic and
elegant, the implementation is a bit magical and may be opaque,
especially to someone who does not sufficiently understand special
Python object protocols.

Good embedded documentation can help users appreciate the magic and not
get scared by it.