bugfix: send CTRL-C through controlling pty on termination
why would you want to send CTRL-C to a process when it terminates?

For example, say we execute a SUID subprocess (e.g., suexec) from a
program and we want to terminate that subprocess.

Unfortunately, since we don't own the process we can't send signals to
it, but somehow when you execute a suid process in a shell and CTRL-C
the process terminates just fine, so how do we do that?

It turns out that you actually can send the process a range of signals
(e.g., SIGINT, SIGQUIT, SIGTSTP) even if you don't own it, but only
through a controlling pty.

What you have to do is give the process its own pty (command.Command
knows how to do that), and then send it a magical VINTR control
character sequence which is intercepted by the kernel's tty device
driver and converted into a SIGINT to all processes in the current
process group for the tty.

Evil poorly documented voodoo that took me hours to find.

The termios(3) man page has a listing of the magical control characters
you can send. VINTR is CTRL-C.

Python has a very low-level interface to termios.

For example::

    import termios
    import popen4 # in sterile-pylib

    p = popen4.Popen4("cat", pty=True)
    control_characters = termios.tcgetattr(p.tochild.fileno())[-1]

    CTRLC = control_characters[termios.VINTR]
    p.tochild.write(CTRLC)